requirements.txt
# marva_face_recognition module
face_recognition
cv2

# marva_speech module
gtts
pygame

# marva_slack module
slackclient